test: refactor test
- All tests are moved to a package tests
- test.py is now a test suite
- coverage is added to requirements
