+++ requirements.txt
@@ -1,4 +1,5 @@
 click==8.1.3
+coverage==6.4.4
 Flask==2.1.2
 Flask-Cors==3.0.10
 itsdangerous==2.1.2
