+++ tests/__init__.py
@@ -1,0 +1,18 @@
+from tests.test import TestMetricsFaker
+from tests.test_hive_event import TestHiveEvent
+from tests.test_hive_event_collection import TestHiveEventCollection
+from tests.test_hive import TestHive
+from tests.test_hive_collection import TestHiveCollection
+# from tests.test_metrics_history import TestMetricsHistory
+# from tests.test_metrics_history_line import TestMetricsHistoryLine
+# from tests.test_hive_metrics_history_line import TestHiveMetricsHistoryLine
+# from tests.test_station_metrics_history_line import TestStationMetricsHistoryLine
+from tests.test_station import TestStation
+from tests.test_station_collection import TestStationCollection
+# from tests.test_monitored_object import TestMonitoredObject
+# from tests.test_faker import TestFaker
+from tests.test_owner import TestOwner
+from tests.test_hive_collection import TestHiveCollection
+from tests.test_faker_hive import TestHiveFaker
+from tests.test_faker_station import TestStationFaker
+
