+++ tests/test_hive_collection.py
